refactor(python): move AgentRegistry from core/ to services/agents/
The Python AgentRegistry class has always been a service-tier concern
(agent lifecycle is a service domain, not a kernel-internal primitive),
but lived in nexus.core.agent_registry for historical reasons. Move it
alongside its peers - agent_warmup, agent_registration, eviction_manager,
agent_provisioning, agent_rpc_service - under nexus.services.agents.

Pure relocation: import paths updated at all 11 call sites (3 services,
1 core resolver, 1 factory wired tier, 6 tests). No behavioral change.
ruff isort fixed import ordering for the moved module; one stray BOM
marker dropped.

Sets up the next commit, where AgentRegistry becomes a thin shim that
delegates state ops to the Rust services::agent_table::AgentTable SSOT.
