+++ src/nexus/services/agents/agent_registry.py
