+++ src/nexus/services/agents/eviction_manager.py
@@ -25,8 +25,8 @@ from nexus.contracts.qos import EVICTION_ORDER, EvictionContext, PressureLevel, 
 
 if TYPE_CHECKING:
     from nexus.contracts.process_types import AgentDescriptor
-    from nexus.core.agent_registry import AgentRegistry
     from nexus.lib.performance_tuning import EvictionTuning
+    from nexus.services.agents.agent_registry import AgentRegistry
     from nexus.services.agents.eviction_policy import EvictionPolicy
     from nexus.services.agents.resource_monitor import ResourceMonitor
